pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyspray"
version = "0.0.1"
authors = [
    {name = "sillysandvich", email = "teamfortress.stupid@gmail.com"}
]
dependencies = [
    "certifi==2024.7.4",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "colorama==0.4.6",
    "idna==3.7",
    "numpy==2.0.1",
    "opencv-python==4.10.0.84",
    "pillow==10.4.0",
    "quicktex==0.2.1",
    "requests==2.32.3",
    "urllib3==2.2.2"
]
license = {file = "LICENSE"}
description = "A CLI interface for quickly making sprays for Team Fortress 2."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
[project.urls]
Homepage = "https://github.com/sillysandvich/pyspray"

